Remove monkey patch and add a shim Provider to hook up
